requirements.txt
briefcase
quart
hypercorn
toga
tortoise-orm
platformdirs